Move search modal views to a separate module.
